Added an option to skip updating the URL
In my particular use of this library I don't want to actually update the url, I just want to scroll to a location on the screen. Changing the URL reloads the page and ruins things. Since the default behavior of the library is to update the URL, the option I've added is one to skip instead of enable.