Fix link navigation
Scroll to link if on same page or allow browser to navigate to it if not on same page. This time should be based off of the Anchor tag no matter what. Previous attempt would receive the tag at the level directly under mouse cursor which was not guaranteed to be the actual Anchor. Could be any tag such as Italics. Uses existing code that checks if the user clicked a smooth scroll link.